+++ tests/testthat/setup.R
@@ -3,9 +3,9 @@ withr::local_options(
   .local_envir = teardown_env()
 )
 
-# use "*" as version for CI and tests
+# use git version of extendr-api for CI and tests
 options(
   "rextendr.extendr_deps" = list(
-    `extendr-api` = "*"
+    `extendr-api` = list(git = "https://github.com/extendr/extendr")
   )
 )